✨ Separate global layout from website layout
